Reduce flickering playlist in PartyMode
Flicker was caused by multiple calls of animated createPlaylist while processing playlist updates in PartyMode. With this change we mainly rely on the recognition of a new PartyMode song which triggers a non-animated playlist update. All other triggers are ignored, this includes TCP notifications or pressing skip buttons.
In addition we only update the playlist in PartyMode, if we are actually on the music playlist. This avoids unintended switching to music playlist while browsing another playlist.