Add support to handle URLs when calling the app
Implement application:openURL:options: and introduce a helper method to map the given name or host to the app's server list and connect to this server. In case there is no matching server in the list, no server will be connected and even the last connected server will be disconnected. This ensures the desired server is connected, no matter what.

Change the URL scheme to "kodiremote", so that the remote app can be called via URL like "kodiremote://host" or "kodiremote://name".